Move stats to top of result page, part 2.
