+++ hypha/apply/funds/templates/funds/submissions_result.html
@@ -17,7 +17,6 @@
     </div>
 </div>
 
-
 <div class="wrapper wrapper--large wrapper--inner-space-medium">
     <div class="wrapper wrapper--bottom-space">
         <h4 class="heading heading--normal">Summery</h4>
